Optimized pdf output, and other small improvements
- pdf output is produced in a similar way as drawing (drawing trapezes
only when it's worth it, and changing line width only when it's worth
it). There's one difference, though: Instead of drawing trapezes and
discs separately, each trapeze is drawn together with two half-discs at
the end (as a single object)

- Moved the threshold for drawing things as constant width into a
constant (LINE_WIDTH_PRECISION); the constant is 1.2 right now, but note
that this means that the error in the linewidth is at most 10%, since if
lines of widths 1.0 and 1.2 are drawn together, they will be drawn in
width 1.1.

- commented out statistics output

- slightly optimized how xo-paint decides whether to draw a trapeze