Improved resizing of small selections
For really thin selections, the area where to click for resizing are shifted outwards, to enlarge the area in which to click for
moving the selection.

It's the function get_resize_sides() which decides what a click at a given point will do. (It's also used to decide about the mouse
cursor.)